solve circular import between infra.py and general.py
