+++ tests/model_serving/model_server/private_endpoint/conftest.py
@@ -11,7 +11,8 @@ from simple_logger.logger import get_logger
 
 from tests.model_serving.model_server.private_endpoint.utils import create_sidecar_pod
 from utilities.constants import KServeDeploymentType, ModelFormat, ModelStoragePath
-from utilities.infra import create_ns, create_isvc
+from utilities.infra import create_isvc
+from utilities.infra import create_ns
 
 LOGGER = get_logger(name=__name__)
