+++ tests/model_serving/model_server/storage/pvc/conftest.py
@@ -11,8 +11,8 @@ from ocp_resources.serving_runtime import ServingRuntime
 from pytest import FixtureRequest
 
 from utilities.constants import KServeDeploymentType
-from utilities.general import download_model_data
-from utilities.infra import get_pods_by_isvc_label, create_isvc
+from utilities.infra import download_model_data, create_isvc
+from utilities.infra import get_pods_by_isvc_label
 
 
 @pytest.fixture(scope="class")
